Module migration_core::commands

The commands exposed by the migration engine core are defined in this module.

Structs

ApplyMigrationsInput

The input to the ApplyMigrations command.

ApplyMigrationsOutput

The output of the ApplyMigrations command.

CreateMigrationInput

The input to the createMigration command.

CreateMigrationOutput

The output of the createMigration command.

DevDiagnosticInput

The devDiagnostic input.

DevDiagnosticOutput

The response type for devDiagnostic.

DiagnoseMigrationHistoryInput

The input to the DiagnoseMigrationHistory command.

DiagnoseMigrationHistoryOutput

The output of the DiagnoseMigrationHistory command.

EvaluateDataLossInput

The input to the evaluateDataLoss command.

EvaluateDataLossOutput

The output of the evaluateDataLoss command.

ListMigrationDirectoriesInput

The input to the ListMigrationDirectories command.

ListMigrationDirectoriesOutput

The output of the ListMigrationDirectories command.

MarkMigrationAppliedInput

The input to the markMigrationApplied command.

MarkMigrationRolledBackInput

The input to the markMigrationRolledBack command.

MigrationFeedback

A data loss warning or an unexecutable migration error, associated with the step that triggered it.

SchemaPushInput

Input to the schemaPush command.

SchemaPushOutput

Output of the schemaPush command.

Enums

DevAction

A suggested action for the CLI migrate dev command.

DriftDiagnostic

A diagnostic returned by diagnoseMigrationHistory when trying to determine whether the development database has the expected schema at its stage in history.

HistoryDiagnostic

A diagnostic returned by diagnoseMigrationHistory when looking at the database migration history in relation to the migrations directory.

Type Definitions

MarkMigrationAppliedOutput

The output of the markMigrationApplied command.

MarkMigrationRolledBackOutput

The output of the markMigrationRolledBack command.