Module migration_core::json_rpc::types

API type definitions used by the methods.

Structs

ApplyMigrationsInput

The input to the applyMigrations command.

ApplyMigrationsOutput

The output of the applyMigrations command.

CreateDatabaseParams

CreateDatabaseResult

CreateMigrationInput

The input to the createMigration command.

CreateMigrationOutput

The output of the creatMigration command.

DatabaseIsBehindFields

DbExecuteParams

The type of params accepted by dbExecute.

DbExecuteResult

The type of results returned by dbExecute.

DebugPanicInput

DebugPanicOutput

DevActionReset

DevDiagnosticInput

The request type for devDiagnostic.

DevDiagnosticOutput

The response type for devDiagnostic.

DiagnoseMigrationHistoryInput

The request params for the diagnoseMigrationHistory method.

DiagnoseMigrationHistoryOutput

The result type for diagnoseMigrationHistory responses.

DiffParams

The type of params for the diff method.

DiffResult

The result type for the diff method.

EnsureConnectionValidityParams

EnsureConnectionValidityResult

EvaluateDataLossInput

The input to the evaluateDataLoss command.

EvaluateDataLossOutput

The output of the evaluateDataLoss command.

GetDatabaseVersionInput

GetDatabaseVersionOutput

ListMigrationDirectoriesInput

The input to the listMigrationDirectories command.

ListMigrationDirectoriesOutput

MarkMigrationAppliedInput

The names of the migrations in the migration directory. Empty if no migrations are found.

MarkMigrationAppliedOutput

The output of the markMigrationApplied command.

MarkMigrationRolledBackInput

The input to the markMigrationRolledBack command.

MarkMigrationRolledBackOutput

The output of the markMigrationRolledBack command.

MigrationFeedback

A data loss warning or an unexecutable migration error, associated with the step that triggered it.

PathContainer

ResetInput

ResetOutput

SchemaContainer

An object with a schema field.

SchemaPushInput

Request params for the schemaPush method.

SchemaPushOutput

Response result for the schemaPush method.

UrlContainer

An object with a url field.

Enums

DatasourceParam

The path to a live database taken as input. For flexibility, this can be the path to a Prisma schema file containing the datasource, or the whole Prisma schema as a string, or only the connection string. See variants.

DbExecuteDatasourceType

The location of the live database to connect to.

DevAction

A suggested action for the CLI migrate dev command.

DiffTarget

A supported source for a database schema to diff in the diff command.

HistoryDiagnostic

A diagnostic returned by diagnoseMigrationHistory when looking at the database migration history in relation to the migrations directory.